Pull code into a new file
